+++ Src/CSharpier.VSCode/src/CSharpierProcessProvider.ts
@@ -13,6 +13,7 @@ import { execDotNet } from "./DotNetProvider";
 import { NullCSharpierProcess } from "./NullCSharpierProcess";
 import { CSharpierProcessServer } from "./CSharpierProcessServer";
 import { ICSharpierProcess2 } from "./ICSharpierProcess";
+import { runFunctionsUntilResultFound } from "./RunFunctionsUntilResultFound";
 
 export class CSharpierProcessProvider implements Disposable {
     warnedForOldVersion = false;
@@ -37,7 +38,8 @@ export class CSharpierProcessProvider implements Disposable {
         );
 
         this.disableCSharpierServer =
-            workspace.getConfiguration("csharpier").get<boolean>("dev.disableCSharpierServer") ?? false;
+            workspace.getConfiguration("csharpier").get<boolean>("dev.disableCSharpierServer") ??
+            false;
 
         window.onDidChangeActiveTextEditor((event: TextEditor | undefined) => {
             if (event?.document?.languageId !== "csharp") {
@@ -114,6 +116,51 @@ export class CSharpierProcessProvider implements Disposable {
     };
 
     private getCSharpierVersion = (directoryThatContainsFile: string): string => {
+        const csharpierVersion = runFunctionsUntilResultFound(
+            () => this.findVersionInCsProjOfParentsDirectories(directoryThatContainsFile),
+            () => this.findCSharpierVersionInToolOutput(directoryThatContainsFile, false),
+            () => this.findCSharpierVersionInToolOutput(directoryThatContainsFile, true),
+        );
+
+        if (!csharpierVersion) {
+            return "";
+        }
+
+        const versionWithoutHash = csharpierVersion.split("+")[0];
+        this.logger.debug(`Using ${versionWithoutHash} as the version number.`);
+        return versionWithoutHash;
+    };
+
+    private findCSharpierVersionInToolOutput = (
+        directoryThatContainsFile: string,
+        isGlobal: boolean,
+    ) => {
+        const command = `tool list${isGlobal ? " -g" : ""}`;
+        const output = execDotNet(command, directoryThatContainsFile).toString().trim();
+
+        this.logger.debug(`Running 'dotnet ${command}' to look for version`);
+        this.logger.debug(`Output was: \n${output}`);
+
+        const lines = output
+            .split("\n")
+            .map(line => line.trim())
+            .filter(line => line.length > 0);
+
+        // The first two lines are headers
+        for (let i = 2; i < lines.length; i++) {
+            const columns = lines[i].split(/\s{2,}/);
+            if (columns.length >= 2) {
+                if (columns[0].toLowerCase() === "csharpier") {
+                    return columns[1];
+                }
+            }
+        }
+    };
+
+    private findVersionInCsProjOfParentsDirectories = (directoryThatContainsFile: string) => {
+        this.logger.debug(
+            `Looking for csproj in or above ${directoryThatContainsFile} that references CSharpier.MsBuild`,
+        );
         let currentDirectory = directoryThatContainsFile;
         let parentNumber = 0;
         while (parentNumber < 30) {
@@ -122,45 +169,12 @@ export class CSharpierProcessProvider implements Disposable {
                 return csProjVersion;
             }
 
-            const dotnetToolsPath = path.join(currentDirectory, ".config/dotnet-tools.json");
-            this.logger.debug(`Looking for ${dotnetToolsPath}`);
-            if (fs.existsSync(dotnetToolsPath)) {
-                const data = JSON.parse(fs.readFileSync(dotnetToolsPath).toString());
-                const version = data.tools.csharpier?.version;
-                if (version) {
-                    this.logger.debug("Found version " + version + " in " + dotnetToolsPath);
-                    return version;
-                }
-            }
-
             const nextDirectory = path.join(currentDirectory, "..");
             if (nextDirectory === currentDirectory) {
                 break;
             }
             currentDirectory = nextDirectory;
             parentNumber++;
-        }
-
-        this.logger.debug(
-            "Unable to find dotnet-tools.json, falling back to running dotnet csharpier --version",
-        );
-
-        let outputFromCsharpier: string;
-
-        try {
-            outputFromCsharpier = execDotNet(`csharpier --version`, directoryThatContainsFile)
-                .toString()
-                .trim();
-
-            this.logger.debug(`dotnet csharpier --version output: ${outputFromCsharpier}`);
-            const versionWithoutHash = outputFromCsharpier.split("+")[0];
-            this.logger.debug(`Using ${versionWithoutHash} as the version number.`);
-            return versionWithoutHash;
-        } catch (error: any) {
-            const message = !error.stderr ? error.toString() : error.stderr.toString();
-
-            this.logger.debug("dotnet csharpier --version failed with " + message);
-            return "";
         }
     };
 
@@ -240,13 +254,9 @@ export class CSharpierProcessProvider implements Disposable {
                     version,
                 );
             } else if (semver.gte(version, "0.12.0")) {
-                if (
-                    semver.gte(version, serverVersion)
-                    && this.disableCSharpierServer
-                )
-                {
+                if (semver.gte(version, serverVersion) && this.disableCSharpierServer) {
                     this.logger.debug(
-                        "CSharpier server is disabled, falling back to piping via stdin"
+                        "CSharpier server is disabled, falling back to piping via stdin",
                     );
                 }